LPS-61952 Rename com.liferay.portlet.layoutsadmin.util to com.liferay.layouts.admin.kernel.util
