+++ portal-service/src/com/liferay/layouts/admin/kernel/util/packageinfo
